requirements.txt
PyQt5==5.15.10
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pyod==1.1.0
psutil==5.9.5
cryptography==41.0.3
pyyaml==6.0.1
reportlab==4.0.4
yara-python==4.3.1
pynput==1.7.6
watchdog==3.0.0
plyer==2.1.0
matplotlib==3.7.2
seaborn==0.12.2
sqlalchemy==2.0.19
schedule==1.2.0
argparse==1.4.0
